requirements.txt
streamlit
python-dotenv
langchain
langchain-community
faiss-cpu
sentence-transformers
requests
